Some changes

+++ distance/entity/Result.cs
@@ -8,6 +8,6 @@ namespace distance.entity
         public DateTime currentDate { get; set; }
         public int uniqueVisits { get; set; }
         //public int duration { get; set; }
-        public int length { get; set; }
+        public float length { get; set; }
     }
 }
